// types.ts
export interface Database {
  public: {
    Tables: {
      admins: {
        Row: {
          id: string;
          created_at: string;
          user_id: string;
        };
        Insert: {
          id?: string;
          created_at?: string;
          user_id?: string;
        };
        Update: {
          id?: string;
          created_at?: string;
          user_id?: string;
        };
      };
      categories: {
        Row: {
          id: string;
          created_at: string;
          title: string;
          slug: string;
          image: string;
        };
        Insert: {
          id?: string;
          created_at?: string;
          title: string;
          slug: string;
          image?: string;
        };
        Update: {
          id?: string;
          created_at?: string;
          title?: string;
          slug?: string;
          image?: string;
        };
      };
// ...
      canvas_variants: {
        Row: {
          id: string;
          created_at: string;
          title: string;
          options: {
            size: string;
            depth: string;
          };
          placeholders: {
            position: string;
            height: number;
            width: number;
          }[];
          cost: number;
        };
        Insert: {
          id: string;
          created_at?: string;
          title?: string;
          options?: {
            size: string;
            depth: string;
          };
          placeholders?: {
            position: string;
            height: number;
            width: number;
          }[];
          cost?: number;
        };
        Update: {
          id: string;
          created_at?: string;
          title?: string;
          options?: {
            size: string;
            depth: string;
          };
          placeholders?: {
            position: string;
            height: number;
            width: number;
          }[];
          cost?: number;
        };
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      get_total_paid_orders: {
        Args: Record<PropertyKey, never>;
        Returns: number;
      };
      is_admin: {
        Args: Record<PropertyKey, never>;
        Returns: boolean;
      };
    };
    Enums: {
      [_ in never]: never;
    };
  };
}